refactor: user stats distribution
